Fix cast warnings

+++ rzauctionmonitor/parser/monitor/AuctionParser.cpp
@@ -57,7 +57,7 @@ void AuctionParser::onScandir(uv_fs_t* req) {
 		thisInstance->log(LL_Error,
 		                  "Failed to scan dir \"%s\", error: %s(%d)\n",
 		                  req->path,
-		                  uv_strerror(req->result),
+		                  uv_strerror(static_cast<int>(req->result)),
 		                  (int) req->result);
 		return;
 	}
